Only read PCI OptionROM firmware when devices are manually unlocked
This allows us to do two things:

 * Start up much faster, now typically under 150ms on fast machines
 * Stop hard-locking machines with dodgy hardware or non-free kernel drivers

The last point in particular was causing gnome-software a bad reputation, as
fwupd was only auto-started the when gnome-software was launched, which read
the Option ROM which then hardlocked some machines.

Note: We still unconditionally read the Option ROM when verifying, although
this is a seldom used feature, and easier to correlate with hard lockups on
faulty hardware.