Add initial support for showing a dock is connected.
This triggers off of the USB NIC contained in the dock.
It uses a DACI call to query number of connected docks so it should
rule out any rogue USB NIC's that happen to have the same VID/PID.

NOTE: this version field may not be used and instead individual components
will be updated.  This might mean virtual devices for all dock components
instead.